chat-bot/requirements.txt
flask==3.0.0
python-dotenv==1.0.1
google-generativeai==0.8.2
pymongo==4.8.0
requests==2.32.3


